Keep standard deviation constant at 1

+++ src/rainbow.hh
@@ -25,7 +25,7 @@ public:
 	
 	static Rainbow& getInstance();
 	long unsigned int create_thread();
-	double get_action(const long unsigned int thread_id, const std::vector<double> state);
+	int get_action(const long unsigned int thread_id, const std::vector<double> state);
 	void put_reward(const long unsigned int thread_id, const double reward_throughput, const double reward_delay, const double duration);
 	void finish(const long unsigned int thread_id, size_t actions_to_remove, const double time_difference);
 	void delete_thread(const long unsigned int thread_id);
